feat(UploadProgressBar): Improve code
